Test-slim Phase 2 / fluff Phase D: verify then resolve the dormant primitives
Verified both June 10 'declared, not enforced' findings against current code:

- advice_boundary: ENFORCED. Chat tool calls and POST /api/invocations
  dispatch through core.runtime.dispatch_capability ->
  prompt_runtime.run_prompt_capability -> advice_boundary.check(), raising
  AdviceBoundaryDenied (403 / chat refusal). Tests renamed
  test_phase1_advice_boundary.py -> test_advice_boundary.py (no enforced
  coverage deleted; composition test folded in from the old integration
  file); enforced surface documented in ARCHITECTURE.md.

- state_machine: DORMANT. Only app importer was its own unmounted-able
  HTTP API; trust ceremony has its own CeremonyState logic; frontend never
  calls /api/state-machine. Runtime + registry + API moved to
  backend/contrib/state_machine/ (out of the app import graph, header
  names the v0.2 output-lifecycle revival); routes unmounted from
  app.main. Models/migrations stay (audit reconstruction reads the
  tables). Tests parked in backend/tests/dormant/ (collect_ignore in
  tests/conftest.py; CI shard glob is non-recursive and never saw the
  subdir — boundary file tests/test_phase3_phase4_round2_fixes.py intact).

Line counts: backend/app 33,967 -> 32,405 (-1,562); live tests
29,030 -> 27,714 (-1,316); tests/dormant/ +1,409 parked. Tests collected
847 -> 813 live, zero collection errors; suite green locally bar the three
pre-existing macOS-only sandbox failures (pass on Linux CI).
