+++ backend/contrib/__init__.py
@@ -1,0 +1,5 @@
+"""Dormant primitives parked out of the app import graph.
+
+Nothing under ``app/`` may import from ``contrib``. Each package here
+carries a header naming the roadmap item that revives it.
+"""
